Move redirect_fd from cuda tests up to base testing support
