+++ numba/cuda/testing.py
@@ -1,12 +1,14 @@
 from __future__ import print_function, absolute_import, division
 
 import contextlib
-import io
-import os
 import sys
 
 from numba import config, unittest_support as unittest
-from numba.tests.support import captured_stdout, SerialMixin
+from numba.tests.support import (
+    captured_stdout,
+    SerialMixin,
+    redirect_c_stdout,
+)
 
 
 class CUDATestCase(SerialMixin, unittest.TestCase):
@@ -22,23 +24,6 @@ def skip_on_cudasim(reason):
 
 def skip_unless_cudasim(reason):
     return unittest.skipUnless(config.ENABLE_CUDASIM, reason)
-
-
-@contextlib.contextmanager
-def redirect_fd(fd):
-    """
-    Temporarily redirect *fd* to a pipe's write end and return a file object
-    wrapping the pipe's read end.
-    """
-    save = os.dup(fd)
-    r, w = os.pipe()
-    try:
-        os.dup2(w, fd)
-        yield io.open(r, "r")
-    finally:
-        os.close(w)
-        os.dup2(save, fd)
-        os.close(save)
 
 
 class CUDATextCapture(object):
@@ -75,7 +60,6 @@ def captured_cuda_stdout():
     else:
         # The CUDA runtime writes onto the system stdout
         from numba import cuda
-        fd = sys.__stdout__.fileno()
-        with redirect_fd(fd) as stream:
+        with redirect_c_stdout() as stream:
             yield CUDATextCapture(stream)
             cuda.synchronize()
